Time: 0 ms (100%), Space: 20 MB (10.39%) - LeetHub

+++ 0347-top-k-frequent-elements/0347-top-k-frequent-elements.cpp
@@ -11,7 +11,7 @@ public:
         for(auto &it: mp){
             bucket[it.second].push_back(it.first);
         }
-        for(int i= bucket.size() -1 ; i>=0; i--){
+        for(int i= (int)bucket.size() -1 ; i>=0; i--){
             for(int num : bucket[i]){
                 ans.push_back(num);
